requirements.txt
streamlit==1.28.0
anthropic>=0.25.0
PyPDF2==3.0.1
python-dotenv==1.0.0